requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
cffi==1.15.1
Django==4.2.5
django-filter==23.2
djangorestframework==3.14.0
Markdown==3.4.4
mysqlclient==2.2.0
pycparser==2.21
pytz==2023.3.post1
sqlparse==0.4.4
typing-extensions==4.7.1
tzdata==2023.3
